+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 #Added
 * Added timeVariant prop to datetime picker to support 12-hour format.
+* Added example for time variant on terra-dev site
 
 3.12.0 - (April 2, 2019)
 ------------------