+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -12,6 +12,11 @@ Unreleased
 ### Removed
 * Removed 'onInputFocus' prop in favor of the added 'onFocus' prop.
 
+3.13.0 - (April 9, 2019)
+------------------
+### Changed
+* Minor dependency update
+
 3.12.0 - (April 2, 2019)
 ------------------
 ### Changed